Fix bugs and prettify codes
The nav bar at top for now is somehow extended as you adds "p{}" in large-orders.css. As a result, I change all those boxes in the menu bar from  <p>  to <div>. Also, I change the message from <div> to <p>, which will make more sense in this situation.